requirements.txt
typer[all]==0.12.3
whoosh==2.7.4
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2